+++ mordant/src/commonMain/kotlin/com/github/ajalt/mordant/table/TableDslInstances.kt
@@ -185,7 +185,7 @@ internal class SingleColumnBuilderInstance(
             if (empty) {
                 empty = false
             } else {
-                padding = Padding.of(top = defaultPadding)
+                padding = Padding { top = defaultPadding }
             }
         }
         init()