+++ mordant/src/commonMain/kotlin/com/github/ajalt/mordant/table/TableLayout.kt
@@ -71,7 +71,7 @@ internal class TableLayout(private val table: TableBuilderInstance) {
         }
 
         val borders = getStyle(Borders.ALL) { it.cellBorders }
-        val padding = getStyle(Padding.horizontal(1)) { it.padding }
+        val padding = getStyle(Padding { horizontal = 1 }) { it.padding }
         val whitespace = getStyle(Whitespace.PRE) { it.whitespace }
         val textAlign = getStyle(TextAlign.LEFT) { it.align }
         val verticalAlign = getStyle(VerticalAlign.TOP) { it.verticalAlign }