+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ### Added
 - Implemented `hideInput` for prompts on native targets [(#63)](https://github.com/ajalt/mordant/issues/63)
 - Improve cell-width calculation for emoji sequences like skin tone modifiers. [(#64)](https://github.com/ajalt/mordant/issues/64)
-- `withTopPadding`, `withRightPadding`, `withBottomPadding`, `withLeftPadding` extensions for adding padding to single side of a widget
 
+### Changed
+- Replaced most of the `Padding` constructor and `Widget.withPadding` overloads with a unified builder interface
+- 
 ## 2.0.0-beta7
 ### Added
 - Functionality for reading user input: `Terminal.readLineOrNull`, `Terminal.prompt` and various `Prompt` classes